// src/common/configuration/models/IGridStyleClassNames.ts
export interface IGridStyleClassNames {
  gridClassName: string;
  rowClassName: string;
  cellClassName: string;
  smallClassNamesOneUntilTwelve: string[];
  mdClassNamesFromOneUntilTwelve: string[];
  lgClassNamesFromOneUntilTwelve: string[];
  xlClassNamesFromOneUntilTwelve: string[];
  xxlClassNamesFromOneUntilTwelve: string[];
  xxxlClassNamesFromOneUntilTwelve: string[];
}